+++ src/store/account-balance/getters.js
@@ -1,0 +1,7 @@
+export default {
+  error: (state) => state.error,
+  isLoading: (state) => state.isLoading,
+  balances: (state) => state.balances,
+  cw20Contracts: (state) => state.cw20Contracts,
+  swapContracts: (state) => state.swapContracts,
+};