+++ src/store/account/actions.js
@@ -1,13 +1,6 @@
-import {
-  bank,
-  commercio,
-  cosmwasm,
-  distribution,
-  staking,
-  tx,
-} from '@/apis/http';
+import { bank, commercio, distribution, staking, tx } from '@/apis/http';
 import { ACCOUNT, APIS, CONFIG } from '@/constants';
-import { bech32Manager, stringEncoder } from '@/utils';
+import { bech32Manager } from '@/utils';
 
 export default {
   async initAccountDashboard({ commit, dispatch }, address) {
@@ -135,79 +128,4 @@ export default {
   async addTransactions({ dispatch }, { address, offset }) {
     await dispatch('fetchTransactions', { address, offset });
   },
-
-  async initAccountBalance({ commit, dispatch }, address) {
-    commit('reset');
-    commit('setLoading', true);
-    const requests = [
-      dispatch('fetchBalances', address),
-      dispatch('fetchMembership', address),
-      dispatch('fetchWasmBalances', address),
-    ];
-    await Promise.all(requests);
-    commit('setLoading', false);
-  },
-
-  async fetchWasmBalances({ commit }, address) {
-    const balances = [];
-    try {
-      const contracts = await _fetchContracts();
-      if (!contracts.length) {
-        commit('setAllBalances', balances);
-        return;
-      }
-      const all = await _fetchContractsBalance(address, contracts);
-      for (const el of all) {
-        const entries = await _fetchContractHistory(el.contract);
-        if (!entries.length) {
-          continue;
-        }
-        el.name = entries[0].msg.name;
-        el.decimals = entries[0].msg.decimals;
-        el.symbol = entries[0].msg.symbol;
-        balances.push(el);
-      }
-      commit('setAllBalances', balances);
-    } catch (error) {
-      commit('setAllBalances', balances);
-    }
-  },
-};
-
-const _fetchContracts = async () => {
-  try {
-    const response = await cosmwasm.requestContracts(CONFIG.WASM_CODE_ID);
-    return response.data.contracts;
-  } catch (error) {
-    return [];
-  }
-};
-
-const _fetchContractsBalance = async (address, contracts) => {
-  const all = [];
-  try {
-    const obj = { balance: { address } };
-    const query_data = stringEncoder.encodeObjToBase64(obj);
-    for (const contract of contracts) {
-      const response = await cosmwasm.requestContractBalance(
-        contract,
-        query_data
-      );
-      if (parseFloat(response.data.data.balance) > 0) {
-        all.push({ contract, balance: response.data.data.balance });
-      }
-    }
-    return all;
-  } catch (error) {
-    return all;
-  }
-};
-
-const _fetchContractHistory = async (address) => {
-  try {
-    const response = await cosmwasm.requestContractHistory(address);
-    return response.data.entries;
-  } catch (error) {
-    return [];
-  }
 };